merge

+++ src/main/java/com/renta/renta/business/dtos/requests/reservation/ReservationCreateRequest.java
@@ -10,7 +10,7 @@ import java.time.LocalDate;
 @Setter
 public class ReservationCreateRequest implements Serializable {
 
-    private Long carId;
+    private Integer carId;
 
     private LocalDate startDate;
